+++ generator/integration_tests/golden/golden_client_test.cc
@@ -70,7 +70,7 @@ TEST(GoldenClientTest, ListDatabases) {
                                               ListDatabasesRequest const &r) {
         EXPECT_EQ(expected_instance, r.parent());
 
-        return ListDatabasesRange(
+        return google::cloud::internal::MakePaginationRange<ListDatabasesRange>(
             ::google::test::admin::database::v1::ListDatabasesRequest{},
             [](::google::test::admin::database::v1::ListDatabasesRequest const
                    &) {
@@ -497,7 +497,7 @@ TEST(GoldenClientTest, ListBackups) {
       .WillRepeatedly([expected_instance](::google::test::admin::database::v1::
                                               ListBackupsRequest const &r) {
         EXPECT_EQ(expected_instance, r.parent());
-        return ListBackupsRange(
+        return google::cloud::internal::MakePaginationRange<ListBackupsRange>(
             ::google::test::admin::database::v1::ListBackupsRequest{},
             [](::google::test::admin::database::v1::ListBackupsRequest const
                    &) {
@@ -577,7 +577,7 @@ TEST(GoldenClientTest, ListDatabaseOperations) {
           [expected_instance](::google::test::admin::database::v1::
                                   ListDatabaseOperationsRequest const &r) {
             EXPECT_EQ(expected_instance, r.parent());
-            return ListDatabaseOperationsRange(
+            return google::cloud::internal::MakePaginationRange<ListDatabaseOperationsRange>(
                 ::google::test::admin::database::v1::
                     ListDatabaseOperationsRequest{},
                 [](::google::test::admin::database::v1::
@@ -614,7 +614,7 @@ TEST(GoldenClientTest, ListBackupOperations) {
                           ::google::test::admin::database::v1::
                               ListBackupOperationsRequest const &r) {
         EXPECT_EQ(expected_instance, r.parent());
-        return ListBackupOperationsRange(
+        return google::cloud::internal::MakePaginationRange<ListBackupOperationsRange>(
             ::google::test::admin::database::v1::ListBackupOperationsRequest{},
             [](::google::test::admin::database::v1::
                    ListBackupOperationsRequest const &) {