+++ CHANGELOG.md
@@ -9,6 +9,28 @@
 
 ## v1.21.0 - TBD
 
+**BREAKING CHANGES:**
+* Some "Range" types used in the Pub/Sub and Spanner APIs lost a constructor
+  that was never intended to be part of their public APIs. Users who were not
+  directly constructing these ranges will not be affected. All uses of these
+  types that was _intended_ to be public will continue to work, i.e., they will
+  all continue to work as ranges in for loops, etc. The affected types are:
+  * `google/cloud/pubsub/subscription_admin_connection.h`
+    * `using ListSubscriptionsRange = google::cloud::internal::PaginationRange`
+    * `using ListSnapshotsRange = google::cloud::internal::PaginationRange`
+  * `google/cloud/pubsub/topic_admin_connection.h`
+    * `using ListTopicsRange = google::cloud::internal::PaginationRange`
+    * `using ListTopicSubscriptionsRange = google::cloud::internal::PaginationRange`
+    * `using ListTopicSnapshotsRange = google::cloud::internal::PaginationRange`
+  * `google/cloud/spanner/database_admin_connection.h`
+    * `using ListDatabaseRange = google::cloud::internal::PaginationRange`
+    * `using ListBackupOperationsRange = google::cloud::internal::PaginationRange`
+    * `using ListDatabaseOperationsRange = google::cloud::internal::PaginationRange`
+    * `using ListBackupsRange = google::cloud::internal::PaginationRange`
+  * `google/cloud/spanner/instance_admin_connection.h`
+    * `using ListInstancesRange = google::cloud::internal::PaginationRange`
+    * `using ListInstanceConfigsRange = google::cloud::internal::PaginationRange`
+
 ## v1.20.0 - 2020-11
 
 ### Bigtable